refactor: split agent orchestrator modules
